Better rendering structure + add filesystem cache
